refactor(ui): extract table logic into independent components
